% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{alignGenes}
\alias{alignGenes}
\title{align genes in counts to profiles for other functions}
\usage{
alignGenes(counts, profiles)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{profiles}{Matrix of reference profiles holding mean expression of genes x cell types. 
Input linear-scale expression, with genes in rows and cell types in columns.}
}
\value{
a cells * genes count matrix for shared genes only
}
\description{
Process counts to have genes shared with profiles
}
