% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_anchor_cells.R
\name{choose_anchors_from_stats}
\alias{choose_anchors_from_stats}
\title{Choose anchor cells given anchor stats}
\usage{
choose_anchors_from_stats(
  counts,
  neg = NULL,
  bg,
  anchorstats = NULL,
  cos = NULL,
  llr = NULL,
  n_cells = 500,
  min_cosine = 0.3,
  min_scaled_llr = 0.01,
  insufficient_anchors_thresh = 20
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}

\item{anchorstats}{Output from get_anchor_stats. Must provide either this or both cos and llr matrices.}

\item{cos}{Matrix of cosine distances from reference profiles. Cells in rows, cell types in columns.}

\item{llr}{Matrix of log likelihood ratios from reference profiles. Cells in rows, cell types in columns.}

\item{n_cells}{Up to this many cells will be taken as anchor points}

\item{min_cosine}{Cells must have at least this much cosine similarity to a fixed profile to be used as an anchor}

\item{min_scaled_llr}{Cells must have (log-likelihood ratio / totalcounts) above this threshold to be used as an anchor}

\item{insufficient_anchors_thresh}{Cell types that end up with fewer than this many anchors will be discarded.}
}
\value{
A vector holding anchor cell assignments (or NA) for each cell in the counts matrix
}
\description{
Starting with cosine distances and log likelihood ratios, choose anchor cells.
}
