% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_init_clust.R
\name{choose_init_clust}
\alias{choose_init_clust}
\title{Choose init clusts aware of fixed_profiles}
\usage{
choose_init_clust(
  counts,
  fixed_profiles,
  bg,
  align_genes = TRUE,
  size,
  n_clusts,
  thresh = 0.9
)
}
\arguments{
\item{counts}{Counts matrix from inside insitutype}

\item{fixed_profiles}{Matrix of expression profiles of pre-defined clusters,
e.g. from previous scRNA-seq. These profiles will not be updated by the EM algorithm.
Colnames must all be included in the init_clust variable.}

\item{bg}{Expected background}

\item{size}{The size parameter to assume for the NB distribution.}

\item{n_clusts}{Number of free clusters}

\item{thresh}{Threshold above which cells will be called as well-fit by fixed_profiles and initialized as a fixed_profile.}
}
\value{
A vector of initial cluster assignments
}
\description{
Choose init clusts such that cells with good logliks under fixed_profiles aren't initialized to new clusters
}
