% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseClusterNumber.R
\name{chooseClusterNumber}
\alias{chooseClusterNumber}
\title{Estimate the correct number of clusters using a subset of the data}
\usage{
chooseClusterNumber(
  counts,
  neg,
  bg = NULL,
  anchors = NULL,
  init_clust = NULL,
  n_clusts = 2:12,
  max_iters = 10,
  subset_size = 1000,
  align_genes = TRUE,
  plotresults = FALSE,
  nb_size = 10,
  pct_drop = 0.005,
  min_prob_increase = 0.05,
  ...
)
}
\arguments{
\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}

\item{anchors}{Vector giving "anchor" cell types, for use in semi-supervised clustering. 
Vector elements will be mainly NA's (for non-anchored cells) and cell type names
for cells to be held constant throughout iterations.}

\item{init_clust}{Vector of initial cluster assignments.}

\item{n_clusts}{Vector giving a range of cluster numbers to consider.}

\item{max_iters}{Number of iterations in each clustering attempt. Recommended to choose
a smaller number for a quicker, approximate clustering.}

\item{subset_size}{Number of cells to include in clustering.}

\item{align_genes}{Logical, for whether to align the genes in fixed_profiles with the colnames in count}

\item{plotresults}{Logical, for whether to plot the results.}

\item{nb_size}{The size parameter to assume for the NB distribution.}

\item{...}{Arguments passed to nbclust.}
}
\value{
A list, with the following elements:
\itemize{
 \item
}
}
\description{
For a subset of the data, perform clustering under a range of cluster numbers.
 Report on loglikelihood vs. number of clusters, and suggest a best choice.
}
