% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnbinom_sparse}
\alias{dnbinom_sparse}
\title{negative binomial density function}
\usage{
dnbinom_sparse(x, mu, size_dnb)
}
\arguments{
\item{x}{dgCMatrix of quantiles}

\item{mu}{dgCMatrix containing Mean of the distribution}

\item{size_dnb}{Dispersion parameter}
}
\value{
dgCMatrix of densities
}
\description{
Probability density function of the negative binomial distribution (written in C++)
}
