% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellEMClust.R
\name{Estep_reference}
\alias{Estep_reference}
\title{E step: update reference cell profiles}
\usage{
Estep_reference(counts, clust, neg, fixed_profiles)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{clust}{Vector of cluster assignments, or a matrix of probabilities
of cells (rows) belonging to clusters (columns).}

\item{neg}{Vector of mean background counts}

\item{fixed_profiles}{Matrix of expression profiles of pre-defined clusters,
e.g. from previous scRNA-seq. These profiles will not be updated by the EM algorithm.
Colnames must all be included in the init_clust variable.}
}
\value{
A matrix of cluster profiles, genes * clusters
}
\description{
Given cell assignments (or posterior probabilities), update the pre-defined
 reference cell profiles.
}
