% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbclust.R
\name{Estep_size}
\alias{Estep_size}
\title{Estimate NB size parameter}
\usage{
Estep_size(counts, clust, bg)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{clust}{Vector of cluster assignments, or a matrix of probabilities
of cells (rows) belonging to clusters (columns).}

\item{bg}{Expected background}
}
\value{
A scalar giving the mle for the size parameter
}
\description{
Given cell assignments (or posterior probabilities), update theta.
Assumption is one theta for all data.
}
