% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlatformEffects.R
\name{estimate_platform_effects}
\alias{estimate_platform_effects}
\title{Estimate platform effects / gene scaling factors}
\usage{
estimate_platform_effects(mean_profiles, fixed_profiles)
}
\arguments{
\item{mean_profiles}{Matrix of mean background-subtracted expression of genes * clusters.}

\item{fixed_profiles}{Matrix of pre-specified reference profiles.}
}
\value{
A vector of scaling factors used to multiply the rows of fixed_profiles
}
\description{
Estimates a scaling factor for each gene to bring the fixed_profiles onto the scale of the data.
 Models the log ratio between platforms as a function of gene, cell line, and
 reference expression.
}
