% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{estimateBackground}
\alias{estimateBackground}
\title{Prepare bg data for other functions}
\usage{
estimateBackground(counts, neg, bg = NULL)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}
}
\value{
A named vector for the estimated background of each cell
}
\description{
Process neg data or bg to get background for each cell
}
