% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleProfiles.R
\name{estimatePlatformEffects}
\alias{estimatePlatformEffects}
\title{Platform effect adjustment on reference profiles based on the expression profiles of anchors}
\usage{
estimatePlatformEffects(
  counts,
  neg,
  assay_type = c("rna", "protein"),
  bg = NULL,
  anchors,
  profiles,
  sds = NULL,
  blacklist = NULL
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{assay_type}{Assay type of RNA, protein (default = "rna")}

\item{bg}{Expected background}

\item{anchors}{Vector giving "anchor" cell types, for use in semi-supervised
clustering. Vector elements will be mainly NA's (for non-anchored cells)
and cell type names for cells to be held constant throughout iterations.}

\item{profiles}{Matrix of reference profiles holding mean expression of genes x cell types. 
Input linear-scale expression, with genes in rows and cell types in columns.}

\item{sds}{Matrix of reference profiles holding SDs expression of genes x cell types. 
Input linear-scale expression, with genes in rows and cell types in columns. Only for assay_type of protein}

\item{blacklist}{vector of user-defined genes to be excluded for cell typing (default = NULL)}
}
\value{
A list with five elements: 
\describe{
    \item{rescaled_profiles}{genes * cell types Matrix of rescaled reference profiles with platform effect corrected }
    \item{platformEff_statsDF}{a data.frame for statistics on platform effect estimation with genes in rows and columns for `Gene`, `Beta`, `beta_SE`.}
    \item{anchors}{a named vector of anchors used for platform effect estimation}
    \item{blacklist}{a vector of genes excluded from cell typing, including both outliers identified in platform effect estimation and the user-defined genes}
    \item{lostgenes}{a vector of genes excluded from platform effect estiamtion due to insufficient evidence}
}
}
\description{
The general workflow would be: (1) extract the anchor cells from input; 
(2) Run poisson regression with anchor cells; (3) Filter user defined genes(if any) 
and genes with extreme betas, outside [0.01, 100]; (4) Re-scale Profile with Beta estimates.
}
\details{
Calculates gene-wise scaling factor between reference profiles and the observed profiles of the provided anchors.
}
