% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPCA.R
\name{fastApproxPCA}
\alias{fastApproxPCA}
\title{Fast approximate PCA}
\usage{
fastApproxPCA(mat, ncomp, nsub = 5000)
}
\arguments{
\item{mat}{Regular or sparse Matrix, observations in rows, variables in columns}

\item{ncomp}{Number of components to estimate}

\item{nsub}{Subsample size}
}
\value{
A matrix of n x ncomp approximate PC values
}
\description{
Run fast approximate PCA by deriving components from a subset then applying to all observations
This is implemented to avoid version issues with irlba, which actually runs faster than this.
}
