% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastCohorting.R
\name{fastCohorting}
\alias{fastCohorting}
\title{Quickly split cells into cohorts}
\usage{
fastCohorting(mat, n_cohorts = NULL, gaussian_transform = TRUE)
}
\arguments{
\item{mat}{Matrix of variables to be used in cohorting, cells in rows, and variables in columns.
Recommended to use < 20 variables.}

\item{n_cohorts}{Number of clusters to divide cells into}

\item{gaussian_transform}{Whether to map each variable onto the quantiles of a normal distribution.}
}
\value{
A vector of cohort assignments.
}
\description{
Quickly split cells into cohorts using non-RNA data like spatial context and immunofluorescence values.
Rule of thumb: include any variables that might be informative for cell typing, 
 *except* variables you'll want to analyze later. For example, if you'll later
 perform differential expression as a function of spatial context, then it's 
 safer to exclude spatial context from the cell typing exercise (and therefore 
 from this function).
}
\examples{
data("mini_nsclc")
## simulate immunofluorescence data: 
immunofluordata <- matrix(rpois(n = nrow(mini_nsclc$counts) * 4, lambda = 100), 
                          nrow(mini_nsclc$counts))
cohort <- fastCohorting(immunofluordata, gaussian_transform = TRUE)
table(cohort)
}
