% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_anchor_cells.R
\name{find_anchor_cells}
\alias{find_anchor_cells}
\title{Choose anchor cells}
\usage{
find_anchor_cells(
  counts,
  neg = NULL,
  bg = NULL,
  align_genes = TRUE,
  profiles,
  size = 10,
  n_cells = 500,
  min_cosine = 0.3,
  min_scaled_llr = 0.01,
  insufficient_anchors_thresh = 20
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}

\item{align_genes}{Logical, for whether to align the columns of the counts matrix and the rows of
the profiles matrix based on their names.}

\item{profiles}{Matrix of reference profiles holding mean expression of genes x cell types. 
Input linear-scale expression, with genes in rows and cell types in columns.}

\item{size}{Negative binomial size parameter to be used in loglikelihood calculatoin}

\item{n_cells}{Up to this many cells will be taken as anchor points}

\item{min_cosine}{Cells must have at least this much cosine similarity to a fixed profile to be used as an anchor}

\item{min_scaled_llr}{Cells must have (log-likelihood ratio / totalcounts) above this threshold to be used as an anchor}

\item{insufficient_anchors_thresh}{Cell types that end up with fewer than this many anchors will be discarded.}
}
\value{
A vector holding anchor cell assignments (or NA) for each cell in the counts matrix
}
\description{
Finds cells with very good fits to the reference profiles, and saves these
 cells for use as "anchors" in the semi-supervised learning version of nbclust.
}
