% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightpath_layout.R
\name{flightpath_layout}
\alias{flightpath_layout}
\title{"Flightpath" (umap-like) plot of clustering results}
\usage{
flightpath_layout(probs, cluster_xpos = NULL, cluster_ypos = NULL)
}
\arguments{
\item{probs}{Matrix of cells' probabilities of belonging to each cluster.}

\item{cluster_xpos}{Vector of cluster centroids' x position (i.e. where you want each cell type to appear in the plot)}

\item{cluster_ypos}{Vector of cluster centroids' y position}
}
\value{
A list with two elements:
\enumerate{
\item clustpos: a matrix of cluster centroids * x,y positions in the flightpath plot
\item cellpos: A matrix of cells * x,y positions in the flightpath plot
}
}
\description{
Arrays cells in 2d space based on their probability of belonging to a given cluster.
}
