% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightpath_layout.R
\name{flightpath_plot}
\alias{flightpath_plot}
\title{Plot flightpath results}
\usage{
flightpath_plot(flightpath_result = NULL, insitutype_result = NULL, col = NULL)
}
\arguments{
\item{col}{Optional, a vector of cell colors}

\item{fp}{The list output by the flightpath_layout function. Two elements: clustpos, cellpos and clust}
}
\value{
a ggplot object
}
\description{
Plot flightpath results
}
