% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_profiles_protein.R
\name{gen_profiles_protein_annotation}
\alias{gen_profiles_protein_annotation}
\title{Generate the mean reference profile and its SD reference profile from an annotation file
This function is only for protein data set with known anchor cells and their cell types}
\usage{
gen_profiles_protein_annotation(exp.mat, anno)
}
\arguments{
\item{exp.mat}{a matrix of raw protein expression data. cells are in rows and proteins are in columns}

\item{anno}{a data frame or matrix of cell types for anchor cells or manually annotated cell typing information for some cells. Should include cell_ID and celltype at least.}
}
\value{
A list, with the following elements:
\enumerate{
\item mean.ref.profile: a matrix of cluster-specific expression profiles. proteins * cell types
\item SDs.ref.profile: a matrix of standard deviation profiles of pre-defined clusters. proteins * cell types
\item anchors: a vector giving "anchor" cell types. Vector elements will be mainly NA's (for non-anchored cells)
}
}
\description{
Generate the mean reference profile and its SD reference profile from an annotation file
This function is only for protein data set with known anchor cells and their cell types
}
