% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_profiles_protein.R
\name{gen_profiles_protein_expression}
\alias{gen_profiles_protein_expression}
\title{Generate the mean reference profile and its SD reference profile based on the data itself
This function is based on signature matrix included in CELESTA package 
First, we rebuild a nested cell typing lists based on the 2-D signature matrix
Second, we identify anchor cells ranked by their expression level for each cell type's protein marker
Third, we estimate averaged expression level and SDs for proteins and cell types using the anchors}
\usage{
gen_profiles_protein_expression(
  exp.mat,
  sig_mat = NULL,
  cutoff = 0.9,
  min.num.cells = 30,
  keep_marker_proteins = FALSE
)
}
\arguments{
\item{exp.mat}{a matrix of raw protein expression data. cells are in rows and proteins are in columns}

\item{sig_mat}{a signature matrix of cell types. cell types x protein markers}

\item{cutoff}{a cutoff of quantile. e.g) cutoff=0.9 means that top 90 percentiles of cells are called anchors for the protein expression}

\item{min.num.cells}{a minimum number of cells each cell type to estimate its mean or SDs. default value is 30.}

\item{keep_marker_proteins}{whether just marker proteins from the signature matrix is kept. default value is FALSE, which returns all proteins included in the data}
}
\value{
A list, with the following elements:
\enumerate{
\item mean.ref.profile: a matrix of cluster-specific expression profiles. proteins x cell types
\item SDs.ref.profile: a matrix of standard deviation profiles of pre-defined clusters. proteins x cell types
\item anchors: a vector giving "anchor" cell types. Vector elements will be mainly NA's (for non-anchored cells)
}
}
\description{
Generate the mean reference profile and its SD reference profile based on the data itself
This function is based on signature matrix included in CELESTA package 
First, we rebuild a nested cell typing lists based on the 2-D signature matrix
Second, we identify anchor cells ranked by their expression level for each cell type's protein marker
Third, we estimate averaged expression level and SDs for proteins and cell types using the anchors
}
\examples{
data("tonsil_protein")
data("human_signature")
data("mouse_signature")
references <- gen_profiles_protein_expression(
 exp.mat=tonsil_protein$counts,
 sig_mat=NULL)
}
