% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling functions.R
\name{geoSketch}
\alias{geoSketch}
\title{geoSketch
Function for sampling cells evenly across expression space

Given an expression matrix, evenly bin cells across expression space and return bin labels and/or sampled cellIDs}
\usage{
geoSketch(
  X,
  N,
  alpha = 0.1,
  max_iter = 200,
  returnBins = FALSE,
  minCellsPerBin = 1,
  seed = NULL
)
}
\arguments{
\item{X}{feature matrix with cellIDs as rows and featureIDs as columns (can be counts, normalized expression, PCA, UMAP, etc.)}

\item{N}{desired sample size}

\item{alpha}{defines the acceptable minimum number of bins to sample from as `(1-alpha)*N`}

\item{max_iter}{maximum number of iterations used to achieve an acceptable minimum number of bins}

\item{returnBins}{determines whether or not to pass back bin labels for each cell}

\item{minCellsPerBin}{the minimum number of cells required for a bin to be considered for sampling}

\item{seed}{set seed for random sampling}
}
\value{
sampledCells, a vector of cellIDs sampled using the geometric sketching method

Plaid, a named vector of binIDs where names correspond to cellIDs
}
\description{
Function for creating a biased sample of a given dataset with the aim of retaining cells with unique expression vectors
}
\examples{

# Example use of geometric sketching to return sampled list of cellIDs and binIDs
library(Ptolemy) # Load library
data(mini_tma) # Load data

mini_tma <- runPCA(mini_tma, expression_values="raw", ncp = 5) # Run PCA
X <- Giotto:::select_dimReduction(mini_tma, name = "pca") # Use PCA results as expression matrix
N <- 100 # Select the desired sample size

sampledCells <- geoSketch(X, N) # Generate list of sampled cells
geoSketchRes <- geoSketch(X, N, returnBins=TRUE) # Generate list of sampled cells and return named vector of binIDs used for sampling
}
