% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoSketch.R
\name{geoSketch_get_plaid}
\alias{geoSketch_get_plaid}
\title{Bin cells into "plaids"}
\usage{
geoSketch_get_plaid(
  X,
  N,
  alpha = 0.1,
  max_iter = 200,
  returnBins = FALSE,
  minCellsPerBin = 1
)
}
\arguments{
\item{X}{feature matrix with cellIDs as rows and featureIDs as columns (can
be counts, normalized expression, PCA, UMAP, etc.)}

\item{N}{desired sample size}

\item{alpha}{defines the acceptable minimum number of bins to sample from as
`(1-alpha)*N`}

\item{max_iter}{maximum number of iterations used to achieve an acceptable
minimum number of bins}

\item{returnBins}{determines whether or not to pass back bin labels for each
cell}

\item{minCellsPerBin}{the minimum number of cells required for a bin to be
considered for sampling}
}
\value{
Plaid, a named vector of binIDs where names correspond to cellIDs
}
\description{
Assign cells to "plaids", very rough clusters.
}
\examples{
data("mini_nsclc")
geoSketch_get_plaid(mini_nsclc$counts, 100)
}
