% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoSketch.R
\name{geoSketch_sample_from_plaids}
\alias{geoSketch_sample_from_plaids}
\title{Subsample from plaids}
\usage{
geoSketch_sample_from_plaids(Plaid, N, seed = NULL)
}
\arguments{
\item{Plaid}{Vector of cells' plaid IDs}

\item{N}{desired sample size}

\item{seed}{set seed for random sampling}
}
\value{
Plaid, a named vector of binIDs where names correspond to cellIDs

sampledCells, a vector of cellIDs sampled using the geometric sketching method
}
\description{
Sample cells, trying to give each plaid equal representation
}
