% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_anchor_cells.R
\name{get_anchor_stats}
\alias{get_anchor_stats}
\title{Get anchor stats}
\usage{
get_anchor_stats(
  counts,
  neg = NULL,
  bg = NULL,
  align_genes = TRUE,
  profiles,
  sds = NULL,
  size = 10,
  assay_type = c("rna", "protein"),
  min_cosine = 0.3
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}

\item{align_genes}{Logical, for whether to align the columns of the counts matrix and the rows of
the profiles matrix based on their names.}

\item{profiles}{Matrix of reference profiles holding mean expression of genes x cell types. 
Input linear-scale expression, with genes in rows and cell types in columns.}

\item{sds}{Matrix of reference profiles holding SDs expression of genes x cell types. 
Input linear-scale expression, with genes in rows and cell types in columns. Only for assay_type of protein}

\item{size}{Negative binomial size parameter to be used in likelihood calculation.}

\item{assay_type}{Assay type of RNA, protein (default = "rna")}

\item{min_cosine}{Cells must have at least this much cosine similarity to a fixed profile to be used as an anchor.}
}
\value{
A list with two elements: cos, the matrix of cosine distances;
 and llr, the matrix of log likelihood ratios of each cell under each cell type vs. the 2nd best cell type.
}
\description{
Compute the statistics used in finding anchor cells.
Often the anchor cell selection process will involve some trial-and-error. 
This function performs the computationally-expensive steps that only need to 
happen once.
}
\examples{
data("ioprofiles")
data("mini_nsclc")
get_anchor_stats(counts = mini_nsclc$counts,
                 neg = Matrix::rowMeans(mini_nsclc$neg),
                 profiles = ioprofiles,
                 sds=NULL, 
                 assay_type = "RNA")
}
