% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialContext.R
\name{get_neighborhood_expression}
\alias{get_neighborhood_expression}
\title{Calculate neighborhood expression}
\usage{
get_neighborhood_expression(counts, neighbors)
}
\arguments{
\item{counts}{Single cell expression matrix}

\item{neighbors}{A neighbors adjacency matrix}
}
\value{
A matrix in the same dimensions as \code{counts}, giving the expression profile of each cell's neighborhood.
}
\description{
Calculates the expression profile of each cell's neighborhood
}
