% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightpath_layout.R
\name{getMeanClusterConfidence}
\alias{getMeanClusterConfidence}
\title{Summarize clusters' mean confidence}
\usage{
getMeanClusterConfidence(probs)
}
\arguments{
\item{probs}{Matrix of probabilities}
}
\value{
a vector of mean confidences, with values of 1 corresponding to clusters with only prob == 1
}
\description{
Calculate the mean confidence of the cell calls from each cluster
}
\examples{
data("mini_nsclc")
probs <- sapply(rownames(mini_nsclc$counts), function(x) {a = runif(10); a/sum(a)})
dimnames(probs)[[1]] <- letters[1:10]
probs <- t(probs)
getMeanClusterConfidence(probs)
}
