% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfiles.R
\name{getRNAprofiles}
\alias{getRNAprofiles}
\title{Extract mean background-subtracted profiles of RNA data}
\usage{
getRNAprofiles(x, neg, clust)
}
\arguments{
\item{x}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean background counts (or a single value applied to all cells)}

\item{clust}{Vector of cluster assignments, or a matrix of probabilities
of cells (rows) belonging to clusters (columns).}
}
\value{
A matrix of gene x cell type expression profiles.
}
\description{
Given cell assignments and count data, estimate the mean
 profile of each cluster.
}
