% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialContext.R
\name{getSpatialContext}
\alias{getSpatialContext}
\title{Get the neighborhood expression profile around all cells}
\usage{
getSpatialContext(
  counts,
  xy,
  tissue = NULL,
  N = 50,
  rad = NULL,
  dim_reduce_to = NULL
)
}
\arguments{
\item{counts}{Counts matrix}

\item{xy}{2-column matrix of cells' xy positions}

\item{tissue}{vector of tissue IDs. Used to ensure cells for different tissues are never called neighbors}

\item{N}{number of neighbors to use. Specify this or \code{rad}.}

\item{rad}{radius to use to define neighbors. Specify this or \code{N}.}

\item{dim_reduce_to}{If entered, the neighborhood matrix will be reduced to this many PCs}
}
\value{
A matrix of neighborhood expression, potentially by gene, or else by PCs if \code{dim_reduce_to} was set.
}
\description{
Get the neighborhood expression profile around all cells
}
