% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitutype.R
\name{insitutype}
\alias{insitutype}
\alias{insitutype,ANY-method}
\alias{insitutype,SingleCellExperiment-method}
\title{Run insitutype.}
\usage{
insitutype(x, ...)

\S4method{insitutype}{ANY}(
  x,
  neg,
  assay_type,
  bg = NULL,
  anchors = NULL,
  cohort = NULL,
  n_clusts,
  reference_profiles = NULL,
  reference_sds = NULL,
  update_reference_profiles = TRUE,
  sketchingdata = NULL,
  align_genes = TRUE,
  nb_size = 10,
  init_clust = NULL,
  n_starts = 10,
  n_benchmark_cells = 10000,
  n_phase1 = 10000,
  n_phase2 = 20000,
  n_phase3 = 1e+05,
  n_chooseclusternumber = 2000,
  pct_drop = 1/10000,
  min_prob_increase = 0.05,
  max_iters = 40,
  n_anchor_cells = 2000,
  min_anchor_cosine = 0.3,
  min_anchor_llr = 0.03,
  insufficient_anchors_thresh = 20,
  refinement = FALSE,
  rescale = FALSE,
  refit = TRUE
)

\S4method{insitutype}{SingleCellExperiment}(x, ..., assay.type = "counts")
}
\arguments{
\item{x}{Counts matrix (or dgCMatrix), cells * genes.

  Alternatively, a \linkS4class{SingleCellExperiment} object containing such
  a matrix.}

\item{...}{For the \linkS4class{SingleCellExperiment} method, additional
arguments to pass to the ANY method.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{assay_type}{Assay type of RNA, protein}

\item{bg}{Expected background}

\item{anchors}{Vector giving "anchor" cell types, for use in semi-supervised
clustering. Vector elements will be mainly NA's (for non-anchored cells)
and cell type names for cells to be held constant throughout iterations.}

\item{cohort}{Vector of cells' cohort memberships}

\item{n_clusts}{Number of clusters, in addition to any pre-specified cell
types. Enter 0 to run purely supervised cell typing from fixed profiles.
Enter a range of integers to automatically select the optimal number of
clusters.}

\item{reference_profiles}{Matrix of mean expression profiles of pre-defined
clusters, e.g. from previous scRNA-seq. These profiles will not be updated
by the EM algorithm. Columns must all be included in the init_clust
variable.}

\item{reference_sds}{Matrix of standard deviation profiles of pre-defined
clusters. These SD profiles also will not be updated by the EM algorithm. 
Columns must all be included in the init_clust variable. This parameter should
be defined if assay_type is protein. Default is NULL.}

\item{update_reference_profiles}{Logical, for whether to use the data to
update the reference profiles. Default and strong recommendation is TRUE.
(However, if the reference profiles are from the same platform as the
study, then FALSE could be better.)}

\item{sketchingdata}{Optional matrix of data for use in non-random sampling
via "sketching". If not provided, then the data's first 20 PCs will be
used.}

\item{align_genes}{Logical, for whether to align the counts matrix and the
fixed_profiles by gene ID.}

\item{nb_size}{The size parameter to assume for the NB distribution. This 
parameter is only for RNA.}

\item{init_clust}{Vector of initial cluster assignments. If NULL, initial
assignments will be automatically inferred.}

\item{n_starts}{the number of iterations}

\item{n_benchmark_cells}{the number of cells for benchmarking}

\item{n_phase1}{Subsample size for phase 1 (random starts)}

\item{n_phase2}{Subsample size for phase 2 (refining in a larger subset)}

\item{n_phase3}{Subsample size for phase 3 (getting final solution in a very
large subset)}

\item{n_chooseclusternumber}{Subsample size for choosing an optimal number of
clusters}

\item{pct_drop}{the decrease in percentage of cell types with a valid
switchover to another cell type compared to the last iteration. Default
value: 1/10000. A valid switchover is only applicable when a cell has
changed the assigned cell type with its highest cell type probability
increased by min_prob_increase.}

\item{min_prob_increase}{the threshold of probability used to determine a
valid cell type switchover}

\item{max_iters}{Maximum number of iterations.}

\item{n_anchor_cells}{For semi-supervised learning. Maximum number of anchor
cells to use for each cell type.}

\item{min_anchor_cosine}{For semi-supervised learning. Cells must have at
least this much cosine similarity to a fixed profile to be used as an
anchor.}

\item{min_anchor_llr}{For semi-supervised learning. Cells must have
(log-likelihood ratio / totalcounts) above this threshold to be used as an
anchor}

\item{insufficient_anchors_thresh}{Cell types that end up with fewer than
this many anchors after anchor selection will be discarded.}

\item{refinement}{Logical, flag for further anchor refinement, used when update_reference_profiles = TRUE (default = FALSE)}

\item{rescale}{Logical, flag for platform effect correction, used when update_reference_profiles = TRUE (default = FALSE)}

\item{refit}{Logical, flag for fitting reference profiles to anchors, used when update_reference_profiles = TRUE (default = TRUE)}

\item{assay.type}{A string specifying which assay values to use.}
}
\value{
A list, with the following elements: \enumerate{ \item clust: a
  vector given cells' cluster assignments \item prob: a vector giving the
  confidence in each cell's cluster \item logliks: Matrix of cells'
  log-likelihoods under each cluster. Cells in rows, clusters in columns.
  \item profiles: a matrix of cluster-specific expression profiles \item
  anchors: from semi-supervised clustering: a vector giving the identifies
  and cell types of anchor cells }
}
\description{
A wrapper for nbclust, to manage subsampling and multiple random starts.
}
\examples{
data("mini_nsclc")
unsup <- insitutype(
 x = mini_nsclc$counts,
 neg = Matrix::rowMeans(mini_nsclc$neg),
 assay_type = "RNA",
 n_clusts = 8,
 n_phase1 = 200,
 n_phase2 = 500,
 n_phase3 = 2000,
 n_starts = 1,
 max_iters = 5
) # choosing inadvisably low numbers to speed the vignette; using the defaults in recommended.
table(unsup$clust)
}
