% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iocolors}
\alias{iocolors}
\title{Default colors for the cell types in the ioprofiles matrix}
\format{
A named vector
}
\usage{
iocolors
}
\description{
A named vector of colors, giving colors for the cell types of the ioprofiles
 matrix.
}
\keyword{datasets}
