% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ioprofiles}
\alias{ioprofiles}
\title{Matrix of immune cell profiles}
\format{
A matrix of 27161 genes x 16 cell types.
}
\usage{
ioprofiles
}
\description{
A matrix of gene * cell type expected expression values
}
\keyword{datasets}
