% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellEMClust.R
\name{ismax}
\alias{ismax}
\title{For a numeric object, return a logical object of whether each element is the max or not.}
\usage{
ismax(x)
}
\description{
For a numeric object, return a logical object of whether each element is the max or not.
}
