% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbclust.R
\name{lldist}
\alias{lldist}
\title{Calculate the likelihood of the expression mat
  using the reference profiles of x}
\usage{
lldist(x, mat, bg = 0.01, size = 10, digits = 2)
}
\arguments{
\item{x}{matrix of reference cell types, genes x profiles}

\item{mat}{a matrix of expression levels in all cells, cells x genes}

\item{bg}{background level (default: 0.01), usually vector length equal to
number of cells}

\item{size}{the parameters for dnbinom function (default: 10)}

\item{digits}{the number of digits for rounding}
}
\value{
cells x profiles matrix of log likelihoods
}
\description{
Calculate the likelihood of the expression mat
  using the reference profiles of x
}
\examples{
data("mini_nsclc")
data("ioprofiles")
bg <- Matrix::rowMeans(mini_nsclc$neg)
genes <- intersect(dimnames(mini_nsclc$counts)[[2]], dimnames(ioprofiles)[[1]])
mat <- mini_nsclc$counts[, genes]
x <- ioprofiles[genes, ]
lldist(x = x, mat = mini_nsclc$counts, bg = Matrix::rowMeans(mini_nsclc$neg))
}
