% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lls}
\alias{lls}
\title{sum from negative binomial density function}
\usage{
lls(mat, s, x, bg, size_dnb)
}
\arguments{
\item{mat}{dgCMatrix expression counts}

\item{s}{numeric scaling factor}

\item{x}{numeric expression for reference profile}

\item{bg}{numeric background level}

\item{size_dnb}{int Dispersion parameter}
}
\value{
rowSums for matrix of densities
}
\description{
Probability density function of the negative binomial distribution (written in C++)
}
