% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lls_protein}
\alias{lls_protein}
\title{sum from Gaussian density function}
\usage{
lls_protein(mat, bgsub, x, xsd)
}
\arguments{
\item{mat}{dgCMatrix expression matrix}

\item{bgsub}{vector of background expression per cell}

\item{x}{numeric expression for reference profiles}

\item{xsd}{numeric expression for reference SD profiles}
}
\value{
rowSums for matrix of densities
}
\description{
Probability density function of the Gaussian distribution (written in C++)
}
