% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lls_rna}
\alias{lls_rna}
\title{sum from negative binomial density function}
\usage{
lls_rna(mat, bgsub, x, bg, size_dnb)
}
\arguments{
\item{mat}{dgCMatrix expression counts}

\item{bgsub}{vector of background expression per cell}

\item{x}{numeric expression for reference profiles}

\item{bg}{numeric background level}

\item{size_dnb}{int Dispersion parameter}
}
\value{
rowSums for matrix of densities
}
\description{
Probability density function of the negative binomial distribution (written in C++)
}
