% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCells.R
\name{logliks2probs}
\alias{logliks2probs}
\title{convert logliks to probabilities}
\usage{
logliks2probs(logliks)
}
\arguments{
\item{logliks}{Matrix of loglikelihoods, as output by insitytupe. Cells in rows, clusters in columns.}
}
\value{
A matrix of probabilities, in the same dimensions as logliks.
}
\description{
From cell x cluster log-likelihoods, calculate cell x cluster probabilities
}
