% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellEMClust.R
\name{makeClusterNames}
\alias{makeClusterNames}
\title{makeClusterNames}
\usage{
makeClusterNames(cNames, nClust)
}
\arguments{
\item{cNames}{existing cluster names}

\item{nClust}{number of clusters}
}
\value{
a vector of complete cluster names
}
\description{
makeClusterNames
}
