% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxLikCellType.R
\name{maxLikCellType}
\alias{maxLikCellType}
\title{Classify cells based on reference profiles}
\usage{
maxLikCellType(
  counts,
  neg = NULL,
  bg = NULL,
  fixed_profiles,
  nb_size = 10,
  align_genes = TRUE
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell. Can be provided}

\item{bg}{Expected background}

\item{fixed_profiles}{Matrix of expression profiles of pre-defined clusters,
e.g. from previous scRNA-seq. These profiles will not be updated by the EM algorithm.
Colnames must all be included in the init_clust variable.}

\item{nb_size}{The size parameter to assume for the NB distribution.}

\item{align_genes}{Logical, for whether to align the counts matrix and the fixed_profiles by gene ID.}
}
\value{
A list, with the following elements:
\enumerate{
\item clust: a vector given cells' cluster assignments
\item probs: a matrix of probabilies of all cells (rows) belonging to all clusters (columns)
\item logliks: a matrix of each cell's log-likelihood under each cluster
}
}
\description{
Supervised classification of cells. Each cell is assigned to the cell type 
 under which its observed expression profile is most likely.
}
