% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCells.R
\name{mergeCells}
\alias{mergeCells}
\title{Merge cell types in a clustering result}
\usage{
mergeCells(merges = NULL, to_delete = NULL, probs)
}
\arguments{
\item{merges}{A named vector in which the elements give new cluster names and
the names give old cluster names. OK to omit cell types that aren't being merged.}

\item{to_delete}{Cluster names to delete. All cells assigned to these clusters 
will be reassigned to the next best cluster.}

\item{probs}{Matrix of probabilities}
}
\value{
A list with two elements:
\enumerate{
\item clust: a vector of cluster assignments
\item probs: a matrix of probabilities of all cells (rows) belonging to all clusters (columns)
}
}
\description{
Take a user-defined list of cells types to rename/combine, then re-compute
 cluster assignments and probabilities under the merged cell types.
}
\examples{
# define a "merges" input:
merges <- c("macrophages" = "myeloid", "monocytes" = "myeloid", "mDC" = "myeloid",
             "B-cells" = "lymphoid")
}
