% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_nsclc}
\alias{mini_nsclc}
\title{Small example SMI data from a NSCLC tumor}
\format{
A list with the following elements:
 \itemize{
 \item counts A matrix of raw counts, with cells in rows and genes in columns
 \item counts A matrix of negprobe counts, with cells in rows and negprobes in columns
 \item x x positions
 \item y y position
 \item umap umap projection
 }
}
\usage{
mini_nsclc
}
\description{
A 2000-cell excerpt from a 1000-plex SMI study of a NSCLC tumor.
}
\keyword{datasets}
