% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_signature}
\alias{mouse_signature}
\title{Example mouse marker proteins 
 For inputting \code{into gen_profiles_protein_expression()}}
\format{
A matrix of 11844 cells and 2 columns
}
\usage{
mouse_signature
}
\description{
data frame
}
\keyword{datasets}
