% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialContext.R
\name{neighbor_colSums}
\alias{neighbor_colSums}
\title{for each cell, get the colSums of x over its neighbors:}
\usage{
neighbor_colSums(x, neighbors)
}
\arguments{
\item{x}{A matrix}

\item{neighbors}{A (probably sparse) adjacency matrix}
}
\description{
for each cell, get the colSums of x over its neighbors:
}
