% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbclust.R
\name{numCores}
\alias{numCores}
\title{Get number of cores for parallelized operations}
\usage{
numCores()
}
\value{
number of cores to use for mclapply
}
\description{
Get number of cores for parallelized operations
}
\examples{
numCores()
}
