% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightpath_layout.R
\name{plotCellTypeUmap}
\alias{plotCellTypeUmap}
\title{Generate a probability-based umap from the cell typing results}
\usage{
plotCellTypeUmap(semi)
}
\arguments{
\item{semi}{the cell typing result from the cellEMClust function}
}
\value{
a ggplot object
}
\description{
Generate a probability-based umap from the cell typing results
}
