% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineClusters.R
\name{probs2logliks}
\alias{probs2logliks}
\title{Get a logliks matrix from a probabilities matrix}
\usage{
probs2logliks(probs)
}
\arguments{
\item{probs}{probability matrix}
}
\value{
log-transformed matrix
}
\description{
Get a logliks matrix from a probabilities matrix
}
\examples{
a <- runif(10)
probs2logliks(a/sum(a))
}
