% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialContext.R
\name{radiusBasedGraph}
\alias{radiusBasedGraph}
\title{Create spatial network from neighbors within radius R}
\usage{
radiusBasedGraph(x, y, R, subset = 1)
}
\arguments{
\item{x}{spatial coordinate}

\item{y}{spatial coordinate}

\item{R}{radius}

\item{subset}{same length as x,y (see Details)}
}
\value{
sparse adjacency matrix with distances
}
\description{
For each cell identify neighbors within distance \code{R} in Euclidean space
and create an edge between them in graph structure, optionally subset cells
(see Details).
}
\details{
Edges will only be created for cells that have the same \code{subset} value,
usually the slide column id but could also be a slide plus FOV id to only
create edges within an FOV.
}
