% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_anchor_cells.R
\name{refineAnchors}
\alias{refineAnchors}
\title{Filter anchor candidates via projection of reference profiles to anchor-derived UMAP}
\usage{
refineAnchors(
  counts,
  neg = NULL,
  bg = NULL,
  align_genes = TRUE,
  profiles,
  anchor_candidates,
  nn_cells = 500,
  insufficient_anchors_thresh = 20
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}

\item{align_genes}{Logical, for whether to align the columns of the counts matrix and the rows of
the profiles matrix based on their names.}

\item{profiles}{Matrix of reference profiles holding mean expression of genes x cell types. 
Input linear-scale expression, with genes in rows and cell types in columns.}

\item{anchor_candidates}{Named vector of anchor candidates with cell_ID in name and corresponding cell type in values.}

\item{nn_cells}{Number of top nearest neighbors to the projected reference profiles to be selected as final anchor cells.}

\item{insufficient_anchors_thresh}{Cell types that end up with fewer than this many anchors will be discarded.}
}
\value{
anchors, a named vector for the final anchor cells
}
\description{
Calculates expression UMAP model for anchor candidates, then projects reference 
profiles to the anchor-derived UMAP and select anchor candidates within top 
nearest neighbors of the projected reference profiles of same cell type in the 
UMAP as the final anchor cells.
}
