% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineClusters.R
\name{refineClusters}
\alias{refineClusters}
\title{Merge cell types in a clustering result}
\usage{
refineClusters(
  merges = NULL,
  to_delete = NULL,
  subcluster = NULL,
  logliks,
  counts = NULL,
  neg = NULL,
  bg = NULL,
  cohort = NULL
)
}
\arguments{
\item{merges}{A named vector in which the elements give new cluster names and
the names give old cluster names. OK to omit cell types that aren't being merged.}

\item{to_delete}{Cluster names to delete. All cells assigned to these clusters 
will be reassigned to the next best cluster.}

\item{logliks}{Matrix of log-likelihoods output by insitutype, cells in rows, clusters in columns}
}
\value{
A list with two elements:
\enumerate{
\item clust: a vector of cluster assignments
\item prob: Vector of posterior probabilities for each cell type
\item logliks: a matrix of probabilities of all cells (rows) belonging to all clusters (columns)
}
}
\description{
Take a user-defined list of cells types to rename/combine, then re-compute
 cluster assignments and probabilities under the merged cell types.
}
\examples{
# define a "merges" input:
merges = c("macrophages" = "myeloid", "monocytes" = "myeloid", "mDC" = "myeloid",
             "B-cells" = "lymphoid")
# define clusters to delete:
to_delete =  c("a", "f") 
}
