% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleProfiles.R
\name{rescaleProfiles}
\alias{rescaleProfiles}
\title{Rescale rows of reference profiles based on deconvolution of "bulk" data}
\usage{
rescaleProfiles(
  counts,
  neg,
  reference_profiles,
  align_genes = TRUE,
  nb_size = 10,
  max_rescaling = 5
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell. Can be provided}

\item{reference_profiles}{Matrix of expression profiles of pre-defined clusters,
e.g. from previous scRNA-seq. These profiles will not be updated by the EM algorithm.
Colnames must all be included in the init_clust variable.}

\item{align_genes}{Logical, for whether to align the counts matrix and the reference_profiles by gene ID.}

\item{nb_size}{The size parameter to assume for the NB distribution.}

\item{max_rescaling}{Scaling factors will be truncated above by this value and below by its inverse (at 1/value and value)}
}
\value{
\enumerate{
\item profiles: A profiles matrix with the rows rescaled according to platform effects
\item scaling_factors: A vector of genes' scaling factors (what they were multiplied by when updating the reference profiles). 
}
}
\description{
Estimates platform effects and rescales the reference matrix accordingly
}
