% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialUpdate.R
\name{spatialUpdate}
\alias{spatialUpdate}
\title{Update cell typing results with spatial context or other alternative data}
\usage{
spatialUpdate(
  celltype,
  counts,
  neg,
  cohort = NULL,
  altdata = NULL,
  xy = NULL,
  tissue = NULL,
  nb_size = 10,
  assay_type = c("rna", "protein")
)
}
\arguments{
\item{celltype}{Vector of cell type assignments to be updated}

\item{counts}{Counts matrix (or dgCMatrix), cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{cohort}{Vector of cells' cohort memberships. Output of a spatial clustering algorithm makes for good cohorts.}

\item{altdata}{Matrix of cells' alternative data values}

\item{xy}{2-column matrix of cells' xy positions.}

\item{tissue}{Vector giving cells' tissue IDs. Used to separate tissue with overlapping xy coordinates.}

\item{nb_size}{The size parameter to assume for the NB distribution.}

\item{assay_type}{A string specifying which assay values to use.}
}
\description{
Takes cell typing results, then updates it based on alternative data types, 
e.g. spatial context, morphology, or protein expression. Existing cell typing results are 
put into Insitutype's likelihood framework, which then can use alternative data
as a prior to be updated by the expression data to get a new posterior probability 
of cell type.
Performs this operation by 
\enumerate{
\item deriving cell type profiles using InSituType:::Estep(), 
\item assigning cells to "cohorts" (clusters) derived from their alternative data
\item  Inputing the output of steps (1) and (2) into InSituType::insitutype() to 
 re-calculate cell type. 
}
Paths for using alternative data in priority order (choose one; if multiple are input, only the most downstream option will be used):
\enumerate{
\item Input \code{xy} positions (and possibly \code{tissue}). Then cells will be clustered 
 into cohorts based on the expression pattern of their 50 nearest neighboring cells.
\item Input a matrix of alternative data (\code{altdata}) to be automatically clustered into cohorts. This supersedes 
 the altdata matrix derived from the \code{xy} argument.
\item Input your own \code{cohort} vector. This supersedes the above inputs. 
}
}
