% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tonsil_annotation}
\alias{tonsil_annotation}
\title{Matrix of anchor cells' annotation file
 A matrix including cell_ID and cellType for anchors cells}
\format{
A matrix of 11844 cells and 2 columns
}
\usage{
tonsil_annotation
}
\description{
matrix.
}
\keyword{datasets}
