% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tonsil_protein}
\alias{tonsil_protein}
\title{Small example SMI protein data from a tonsil tissue}
\format{
A list with the following elements:
 \itemize{
 \item counts A matrix of raw counts, with cells in rows and proteins in columns
 \item negs A matrix of IgG counts, with cells in rows and IgGs in columns
 \item xy_coord x and y positions
 \item UMAP umap projection
 }
}
\usage{
tonsil_protein
}
\description{
A 21844-cells excerpt from a 68-plex SMI study of a tonsil tissue.
}
\keyword{datasets}
