% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tonsil_reference_profile}
\alias{tonsil_reference_profile}
\title{Reference profile examples from a tonsil tissue
#'}
\format{
A list with the following elements:
 \itemize{
 \item tonsil_reference_profile A matrix of raw counts, with cells in rows and proteins in columns
 \item counts A matrix of IgG counts, with cells in rows and IgGs in columns
 \item xy_coord x and y positions
 \item UMAP umap projection
 }
}
\usage{
tonsil_reference_profile
}
\description{
Reference profile examples from a tonsil tissue
#'
}
\keyword{datasets}
