% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbclust.R
\name{update_logliks_with_cohort_freqs}
\alias{update_logliks_with_cohort_freqs}
\title{Update logliks based on frequencies}
\usage{
update_logliks_with_cohort_freqs(
  logliks,
  cohort,
  minfreq = 1e-06,
  nbaselinecells = 50
)
}
\arguments{
\item{logliks}{Matrix of cells' (rows) loglikelihoods under clusters (columns)}

\item{cohort}{Vector of cells' cohort memberships}

\item{minfreq}{Minimum frequency to give any cell type in any cohort}

\item{nbaselinecells}{Number of cells from baseline distribution to add to the 
cohort-specific frequencies, thereby shrinking each cohort's data towards the population}
}
\value{
An adjusted logliks matrix
}
\description{
Update logliks based on frequencies
}
