% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleProfiles.R
\name{updateProfilesFromAnchors}
\alias{updateProfilesFromAnchors}
\title{Use anchor cells to update reference profiles, simply by taking the mean
profile of the anchors.}
\usage{
updateProfilesFromAnchors(
  counts,
  neg,
  bg = NULL,
  assay_type = c("rna", "protein"),
  anchors
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell. Can be provided}

\item{bg}{Expected background}

\item{assay_type}{Assay type of RNA, protein (default = "rna")}

\item{anchors}{Vector of anchor assignments}
}
\value{
\enumerate{ 
\item updated_profiles: A mean profiles matrix with the rows rescaled
according to platform effects and individual elements updated further 
\item updated_sds: A mean profiles matrix with the rows rescaled
according to platform effects and individual elements updated further}
}
\description{
Uses anchor cells to estimate platform effects / scaling factors to be
applied to the genes/rows of the reference profile matrix. Then uses Bayesian
math to update the individual elements on X.
}
