% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleProfiles.R
\name{updateReferenceProfiles}
\alias{updateReferenceProfiles}
\title{Update reference profiles}
\usage{
updateReferenceProfiles(
  reference_profiles,
  counts,
  neg,
  bg = NULL,
  nb_size = 10,
  anchors = NULL,
  n_anchor_cells = 2000,
  min_anchor_cosine = 0.3,
  min_anchor_llr = 0.01
)
}
\description{
Update reference profiles using pre-specified anchor cells, or if no anchors are specified, by first choosing anchor cells
}
