% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseClusterNumber.R
\name{chooseClusterNumber}
\alias{chooseClusterNumber}
\title{Estimate the correct number of clusters using a subset of the data}
\usage{
chooseClusterNumber(
  counts,
  neg,
  bg = NULL,
  fixed_profiles = NULL,
  init_clust = NULL,
  n_clusts = 2:12,
  max_iters = 10,
  subset_size = 1000,
  align_genes = TRUE,
  plotresults = FALSE,
  nb_size = 10,
  ...
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}

\item{fixed_profiles}{Matrix of expression profiles of pre-defined clusters,
e.g. from previous scRNA-seq. These profiles will not be updated by the EM algorithm.
Colnames must all be included in the init_clust variable.}

\item{init_clust}{Vector of initial cluster assignments.}

\item{n_clusts}{Vector giving a range of cluster numbers to consider.}

\item{max_iters}{Number of iterations in each clustering attempt. Recommended to choose
a smaller number for a quicker, approximate clustering.}

\item{subset_size}{Number of cells to include in clustering.}

\item{align_genes}{Logical, for whether to align the genes in fixed_profiles with the colnames in count}

\item{plotresults}{Logical, for whether to plot the results.}

\item{nb_size}{The size parameter to assume for the NB distribution.}

\item{...}{Arguments passed to nbclust.}
}
\value{
A list, with the following elements:
\itemize{
 \item
}
}
\description{
For a subset of the data, perform clustering under a range of cluster numbers.
 Report on loglikelihood vs. number of clusters, and suggest a best choice.
}
