% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseClusterNumber.R
\name{chooseClusterNumber}
\alias{chooseClusterNumber}
\title{Estimate the correct number of clusters using a subset of the data}
\usage{
chooseClusterNumber(
  counts,
  neg,
  assay_type = c("rna", "protein"),
  bg = NULL,
  fixed_profiles = NULL,
  fixed_sds = NULL,
  cohort = NULL,
  init_clust = NULL,
  n_clusts = 2:12,
  max_iters = 10,
  subset_size = 1000,
  align_genes = TRUE,
  plotresults = FALSE,
  nb_size = 10,
  pct_drop = 0.005,
  min_prob_increase = 0.05,
  ...
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell (default = "rna")}

\item{assay_type}{Assay type of RNA, protein}

\item{bg}{Expected background}

\item{fixed_profiles}{Matrix of cluster profiles to hold unchanged throughout iterations.}

\item{fixed_sds}{Matrix of SDs expression of genes x cell types,to hold unchanged throughout iterations. Only for assay_type of protein}

\item{cohort}{Vector of cells' cohort assignments.}

\item{init_clust}{Vector of initial cluster assignments.}

\item{n_clusts}{Vector giving a range of cluster numbers to consider.}

\item{max_iters}{Number of iterations in each clustering attempt. Recommended to choose
a smaller number for a quicker, approximate clustering.}

\item{subset_size}{Number of cells to include in clustering.}

\item{align_genes}{Logical, for whether to align the genes in fixed_profiles with the colnames in count}

\item{plotresults}{Logical, for whether to plot the results.}

\item{nb_size}{The size parameter to assume for the NB distribution.}

\item{pct_drop}{the decrease in percentage of cell types with a valid switchover to 
another cell type compared to the last iteration. Default value: 1/10000. A valid 
switchover is only applicable when a cell has changed the assigned cell type with its
highest cell type probability increased by min_prob_increase.}

\item{min_prob_increase}{the threshold of probability used to determine a valid cell 
type switchover}

\item{...}{Arguments passed to nbclust.}
}
\value{
A list, with the following elements:
\itemize{
 \item
}
}
\description{
For a subset of the data, perform clustering under a range of cluster numbers.
 Report on loglikelihood vs. number of clusters, and suggest a best choice.
}
\examples{
data("mini_nsclc")
chooseClusterNumber(mini_nsclc$counts, Matrix::rowMeans(mini_nsclc$neg), assay_type="RNA",
 n_clust = 2:5)
}
