% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbclust.R
\name{Estep}
\alias{Estep}
\title{E step: estimate each cluster's mean profile}
\usage{
Estep(counts, clust, neg, assay_type)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{clust}{Vector of cluster assignments, or a matrix of probabilities
of cells (rows) belonging to clusters (columns).}

\item{neg}{Vector of mean background counts}

\item{assay_type}{Assay type of RNA, protein}
}
\value{
A matrix of cluster profiles, genes * clusters
}
\description{
Given cell assignments (or posterior probabilities), estimate the mean
 profile of each cluster.
}
\examples{
data("ioprofiles")
unsup <- insitutype(
 x = mini_nsclc$counts,
 neg = Matrix::rowMeans(mini_nsclc$neg),
 n_clusts = 8,
 n_phase1 = 200,
 n_phase2 = 500,
 n_phase3 = 2000,
 n_starts = 1,
 max_iters = 5,
 assay_type="RNA",
) # choosing inadvisably low numbers to speed the vignette; using the defaults in recommended.
Estep(counts = mini_nsclc$counts, clust = unsup$clust, neg = Matrix::rowMeans(mini_nsclc$neg), assay_type="RNA")
}
