% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_anchor_cells.R
\name{find_anchor_cells}
\alias{find_anchor_cells}
\title{Choose anchor cells}
\usage{
find_anchor_cells(
  counts,
  neg = NULL,
  bg = NULL,
  align_genes = TRUE,
  profiles,
  sds = NULL,
  size = 10,
  assay_type = c("rna", "protein"),
  n_cells = 500,
  min_cosine = 0.3,
  min_scaled_llr = 0.01,
  insufficient_anchors_thresh = 20,
  refinement = FALSE
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}

\item{align_genes}{Logical, for whether to align the columns of the counts
matrix and the rows of the profiles matrix based on their names.}

\item{profiles}{Matrix of reference profiles holding mean expression of genes
x cell types. Input linear-scale expression, with genes in rows and cell
types in columns.}

\item{sds}{Matrix of reference profiles holding SDs expression of genes x cell types. 
Input linear-scale expression, with genes in rows and cell types in columns. Only for assay_type of protein}

\item{size}{Negative binomial size parameter to be used in likelihood calculation. Only for assay_type of RNA}

\item{assay_type}{Assay type of RNA, protein (default = "rna")}

\item{n_cells}{Up to this many cells will be taken as anchor points}

\item{min_cosine}{Cells must have at least this much cosine similarity to a
fixed profile to be used as an anchor}

\item{min_scaled_llr}{Cells must have (log-likelihood ratio / totalcounts)
above this threshold to be used as an anchor}

\item{insufficient_anchors_thresh}{Cell types that end up with fewer than
this many anchors will be discarded.}

\item{refinement}{flag to further refine the anchors via UMAP projection (default = FALSE)}
}
\value{
A vector holding anchor cell assignments (or NA) for each cell in the
  counts matrix
}
\description{
Finds cells with very good fits to the reference profiles, and saves these
cells for use as "anchors" in the semi-supervised learning version of
nbclust. The function would first pick anchor cell candidates through stats 
and then refine anchors based on umap projection.
}
\examples{
data("ioprofiles")
data("mini_nsclc")
sharedgenes <- intersect(colnames(mini_nsclc$counts), rownames(ioprofiles))
find_anchor_cells(counts = mini_nsclc$counts[, sharedgenes], 
                  assay_type="RNA", 
                  sds=NULL,
                  neg = Matrix::rowMeans(mini_nsclc$neg),
                  profiles = ioprofiles)
}
