% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightpath_layout.R
\name{flightpath_plot}
\alias{flightpath_plot}
\title{Plot flightpath results}
\usage{
flightpath_plot(
  flightpath_result = NULL,
  insitutype_result = NULL,
  col = NULL,
  showclusterconfidence = TRUE
)
}
\arguments{
\item{flightpath_result}{The list output by the flightpath_layout function.
Two elements: clustpos, cellpos. Must provide either this or
insitutype_result.}

\item{insitutype_result}{The list output by insitutype or insitutypeML. Must
provide either this or insitutype_result.}

\item{col}{Optional, a vector of cell colors, with length equal to the number
of individual cells.}

\item{showclusterconfidence}{Logical, for whether to label clusters with the
average posterior probability of the cells within them. Gives a readout of
how distinct a cluster is from the others.}
}
\value{
a ggplot object
}
\description{
Plot flightpath results
}
\examples{
data("ioprofiles")
unsup <- insitutype(
 x = mini_nsclc$counts,
 neg = Matrix::rowMeans(mini_nsclc$neg),
 n_clusts = 8,
 n_phase1 = 200,
 n_phase2 = 500,
 n_phase3 = 2000,
 n_starts = 1,
 max_iters = 5,
 assay_type="RNA"
) # choosing inadvisably low numbers to speed the vignette; using the defaults in recommended.
flightpath_plot(insitutype_result = unsup)
}
