% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitutype.R
\name{insitutype}
\alias{insitutype}
\title{Unsupervised and semi-supervised cell typing of single cell expression data.}
\usage{
insitutype(
  counts,
  neg,
  bg = NULL,
  n_clusts = NULL,
  fixed_profiles = NULL,
  sketchingdata = NULL,
  align_genes = TRUE,
  nb_size = 10,
  method = "CEM",
  init_clust = NULL,
  n_starts = 10,
  n_benchmark_cells = 50000,
  n_phase1 = 5000,
  n_phase2 = 20000,
  n_phase3 = 1e+05,
  pct_drop = 1/10000,
  min_prob_increase = 0.05
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}

\item{n_clusts}{Number of clusters, in addition to any pre-specified cell types.}

\item{fixed_profiles}{Matrix of expression profiles of pre-defined clusters,
e.g. from previous scRNA-seq. These profiles will not be updated by the EM algorithm.
Colnames must all be included in the init_clust variable.}

\item{sketchingdata}{Optional matrix of data for use in non-random sampling via "sketching".
If not provided, then the data's first 20 PCs will be used.}

\item{align_genes}{Logical, for whether to align the counts matrix and the fixed_profiles by gene ID.}

\item{nb_size}{The size parameter to assume for the NB distribution.}

\item{method}{Whether to run a SEM algorithm (points given a probability
of being in each cluster) or a classic EM algorithm (all points wholly
assigned to one cluster).}

\item{init_clust}{Vector of initial cluster assignments.
If NULL, initial assignments will be automatically inferred.}

\item{n_starts}{the number of iterations}

\item{n_benchmark_cells}{the number of cells for benchmarking}

\item{n_phase1}{Subsample size for phase 1 (random starts)}

\item{n_phase2}{Subsample size for phase 2 (refining in a larger subset)}

\item{n_phase3}{Subsample size for phase 3 (getting final solution in a very large subset)}

\item{pct_drop}{the decrease in percentage of cell types with a valid switchover to 
another cell type compared to the last iteration. Default value: 1/10000. A valid 
switchover is only applicable when a cell has changed the assigned cell type with its
highest cell type probability increased by min_prob_increase.}

\item{min_prob_increase}{the threshold of probability used to determine a valid cell 
type switchover}
}
\value{
A list, with the following elements:
\enumerate{
\item cluster: a vector given cells' cluster assignments
\item probs: a matrix of probabilies of all cells (rows) belonging to all clusters (columns)
\item profiles: a matrix of cluster-specific expression profiles
\item pct_changed: how many cells changed class at each step
\item logliks: a matrix of each cell's log-likelihood under each cluster
}
}
\description{
A wrapper for nbclust, to manage subsampling and multiple random starts.
}
