% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intisutype.R
\name{insitutype}
\alias{insitutype}
\title{Unsupervised and semi-supervised clustering of single cell expression data.}
\usage{
insitutype(
  counts,
  tissue = NULL,
  neg,
  bg = NULL,
  anchors = NULL,
  cohort = NULL,
  n_clusts,
  fixed_profiles = NULL,
  sketchingdata = NULL,
  align_genes = TRUE,
  nb_size = 10,
  init_clust = NULL,
  n_starts = 10,
  n_benchmark_cells = 50000,
  n_phase1 = 5000,
  n_phase2 = 20000,
  n_phase3 = 1e+05,
  n_chooseclusternumber = 2000,
  pct_drop = 1/10000,
  min_prob_increase = 0.05,
  max_iters = 40,
  n_anchor_cells = 500,
  min_anchor_cosine = 0.3,
  min_anchor_llr = 0.01,
  insufficient_anchors_thresh = 20,
  anchor_replacement_thresh = 0.75
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{tissue}{Vector of cells' tissue IDs.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}

\item{anchors}{Vector giving "anchor" cell types, for use in semi-supervised clustering. 
Vector elements will be mainly NA's (for non-anchored cells) and cell type names
for cells to be held constant throughout iterations.}

\item{cohort}{Vector of cells' cohort memberships}

\item{n_clusts}{Number of clusters, in addition to any pre-specified cell types.
Enter 0 to run purely supervised cell typing from fixed profiles. 
Enter a range of integers to automatically select the optimal number of clusters.}

\item{fixed_profiles}{Matrix of expression profiles of pre-defined clusters,
e.g. from previous scRNA-seq. These profiles will not be updated by the EM algorithm.
Colnames must all be included in the init_clust variable.}

\item{sketchingdata}{Optional matrix of data for use in non-random sampling via "sketching".
If not provided, then the data's first 20 PCs will be used.}

\item{align_genes}{Logical, for whether to align the counts matrix and the fixed_profiles by gene ID.}

\item{nb_size}{The size parameter to assume for the NB distribution.}

\item{init_clust}{Vector of initial cluster assignments.
If NULL, initial assignments will be automatically inferred.}

\item{n_starts}{the number of iterations}

\item{n_benchmark_cells}{the number of cells for benchmarking}

\item{n_phase1}{Subsample size for phase 1 (random starts)}

\item{n_phase2}{Subsample size for phase 2 (refining in a larger subset)}

\item{n_phase3}{Subsample size for phase 3 (getting final solution in a very large subset)}

\item{n_chooseclusternumber}{Subsample size for choosing an optimal number of clusters}

\item{pct_drop}{the decrease in percentage of cell types with a valid switchover to 
another cell type compared to the last iteration. Default value: 1/10000. A valid 
switchover is only applicable when a cell has changed the assigned cell type with its
highest cell type probability increased by min_prob_increase.}

\item{min_prob_increase}{the threshold of probability used to determine a valid cell 
type switchover}

\item{max_iters}{Maximum number of iterations.}

\item{n_anchor_cells}{For semi-supervised learning. Maximum number of anchor cells to use for each cell type.}

\item{min_anchor_cosine}{For semi-supervised learning. Cells must have at least this much cosine similarity to a fixed profile to be used as an anchor.}

\item{min_anchor_llr}{For semi-supervised learning. Cells must have (log-likelihood ratio / totalcounts) above this threshold to be used as an anchor}

\item{insufficient_anchors_thresh}{Cell types that end up with fewer than this many anchors after anchor selection will be discarded.}

\item{anchor_replacement_thresh}{Threshold for calling whether a cluster has wandered away from its anchor cells. 
Clusters whose anchors are reassigned at this rate or higher will be renamed, and only their anchor cells will be used to define a profile for the cluster.}
}
\value{
A list, with the following elements:
\enumerate{
\item clust: a vector given cells' cluster assignments
\item prob: a vector giving the confidence in each cell's cluster
\item logliks: Matrix of cells' log-likelihoods under each cluster. Cells in rows, clusters in columns.
\item profiles: a matrix of cluster-specific expression profiles
\item anchors: from semi-supervised clustering: a vector giving the identifies and cell types of anchor cells
}
}
\description{
A wrapper for runinsitutype, to coordinate clustering across multiple tissues.
}
