% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbclust.R
\name{ismax}
\alias{ismax}
\title{For a numeric object, return a logical object of whether each element is the max or not.}
\usage{
ismax(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
a vecetor of logical values
}
\description{
For a numeric object, return a logical object of whether each element is the max or not.
}
\examples{
ismax(c(3, 5, 5, 2))
}
