% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbclust.R
\name{lldist}
\alias{lldist}
\title{Calculate the likelihood of the vector x using the reference vector of y}
\usage{
lldist(x, mat, bg = 0.01, size = 10, digits = 2)
}
\arguments{
\item{x}{a vector of a reference cell type}

\item{mat}{a matrix of expression levels in all cells}

\item{bg}{background level (default: 0.01)}

\item{size}{the parameters for dnbinom function (default: 10)}

\item{digits}{the number of digits for rounding}
}
\description{
Calculate the likelihood of the vector x using the reference vector of y
}
