% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineClusters.R
\name{logliks2probs}
\alias{logliks2probs}
\title{convert logliks to probabilities}
\usage{
logliks2probs(logliks)
}
\arguments{
\item{logliks}{Matrix of loglikelihoods, as output by insitytupe. Cells in rows, clusters in columns.}
}
\value{
A matrix of probabilities, in the same dimensions as logliks.
}
\description{
From cell x cluster log-likelihoods, calculate cell x cluster probabilities
}
\examples{
data("mini_nsclc")
unsup <- insitutype(
 x = mini_nsclc$counts,
 neg = Matrix::rowMeans(mini_nsclc$neg),
 n_clusts = 8,
 n_phase1 = 200,
 n_phase2 = 500,
 n_phase3 = 2000,
 n_starts = 1,
 max_iters = 5,
 assay_type="RNA"
) # choosing inadvisably low numbers to speed the vignette; using the defaults in recommended.
logliks2probs(unsup$logliks)

}
