% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbclust.R
\name{Mstep}
\alias{Mstep}
\title{M step}
\usage{
Mstep(
  counts,
  means,
  freq,
  bg = 0.01,
  size = 10,
  digits = 2,
  return_loglik = FALSE
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{means}{Matrix of mean cluster profiles,
with genes in rows and clusters in columns.}

\item{freq}{a vector of cells frequencies summing up equal to 1.}

\item{bg}{Expected background}

\item{size}{NB size parameter}

\item{digits}{Round the output to this many digits (saves memory)}

\item{return_loglik}{If TRUE, logliks will be returned. If FALSE, probabilities will be returned.}
}
\value{
Matrix of probabilities of each cell belonging to each cluster
}
\description{
Compute probability that each cell belongs to a given cluster
}
