% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbclust.R
\name{nbclust}
\alias{nbclust}
\title{Cluster via EM algorithm based on cell logliks}
\usage{
nbclust(
  counts,
  neg,
  bg = NULL,
  fixed_profiles = NULL,
  init_profiles = NULL,
  init_clust = NULL,
  nb_size = 10,
  cohort = NULL,
  pct_drop = 1/10000,
  min_prob_increase = 0.05,
  max_iters = 40,
  logresults = FALSE
)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}

\item{init_profiles}{Matrix of cluster profiles under which to begin iterations.
If NULL, initial assignments will be automatically inferred, using init_clust 
if available, and using random clusters if not.}

\item{init_clust}{Vector of initial cluster assignments.
If NULL, initial assignments will be automatically inferred.}

\item{nb_size}{The size parameter to assume for the NB distribution.}

\item{cohort}{Vector of cells' "cohort" assignments, uses to assess frequencies in each cluster.}

\item{pct_drop}{the decrease in percentage of cell types with a valid switchover to 
another cell type compared to the last iteration. Default value: 1/10000. A valid 
switchover is only applicable when a cell has changed the assigned cell type with its
highest cell type probability increased by min_prob_increase.}

\item{min_prob_increase}{the threshold of probability used to determine a valid cell 
type switchover}

\item{max_iters}{Maximum number of iterations

 @importFrom stats lm}
}
\value{
A list, with the following elements:
\enumerate{
\item probs: a matrix of probabilities of all cells (rows) belonging to all clusters (columns)
\item profiles: a matrix of cluster-specific expression profiles
}
}
\description{
Cluster single cell gene expression data using an EM algorithm.
}
