% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoSketch.R
\name{prepDataForSketching}
\alias{prepDataForSketching}
\title{Prepare data for geoSketch}
\usage{
prepDataForSketching(counts)
}
\arguments{
\item{counts}{Counts matrix}
}
\value{
A matrix of data for geoSketch, with cells in rows and features in columns
}
\description{
Process raw counts data for input into geoSketching.
}
\examples{
data("mini_nsclc")
prepDataForSketching(mini_nsclc$counts)
}
